backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and Storage - Supabase
supabase==2.7.4
postgrest==0.13.2

# File Processing - Production Ready
PyMuPDF==1.23.14
pdfplumber==0.10.3
docx2txt==0.8
python-docx==1.1.0

# NLP - spaCy and NLTK
spacy==3.7.2
nltk==3.8.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# LangChain + LangGraph - LLM Orchestration
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langchain-google-genai==0.0.6
langgraph==0.0.25
langchain-community==0.0.10

# LangSmith - LLM Observability
langsmith==0.0.87

# LLM APIs
openai==1.10.0
anthropic==0.18.1
google-generativeai==0.3.2

# Vector Databases
chromadb==0.4.22
faiss-cpu==1.7.4
pinecone-client==3.0.0

# Text Processing and ML - TF-IDF + Fuzzy Matching
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
sentence-transformers==2.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0

# Additional ML utilities
textdistance==4.6.0
jellyfish==0.11.2

# HTTP and async utilities
httpx==0.26.0
aiofiles==23.2.1
requests==2.31.0
tenacity==8.2.3

# Utilities
uuid==1.30
python-dateutil==2.8.2
pytz==2023.4

# Development and Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Production utilities
gunicorn==21.2.0
redis==5.0.1